Restructure folder and files.
