+++ client/src/api/queryUtils.ts
@@ -1,6 +1,6 @@
 import { API_ENDPOINTS } from "./apiEndpoints";
-import { YEAR_BOUNDS, QueryFilter } from "./types";
-import type { QueryPayload } from "./types";
+import { YEAR_BOUNDS, QueryFilter } from "./types/query";
+import type { QueryPayload } from "./types/query";
 import { checkResponse } from "./errors";
 
 /**
