--- client/src/api/types.ts
@@ -1,159 +1,0 @@
-export type QueryFilter = {
-  column: string;
-  operator: string;
-  value: number | number[] | string;
-};
-
-export const YEAR_BOUNDS = { min: 1900, max: 2100 } as const;
-
-export type DatasetRef = { path: string; alias: string };
-
-export type JoinClause = { left: string; right: string; on: string[] };
-
-export type Aggregation = { alias: string; expression: string };
-
-/**
- * Canonical shape for every POST body sent to the WEPPcloud Query Engine.
- * All API helpers should build this type (or a subset) instead of ad-hoc
- * `Record<string, unknown>` objects.
- */
-export type QueryPayload = {
-  datasets: DatasetRef[];
-  columns: string[];
-  filters?: QueryFilter[];
-  joins?: JoinClause[];
-  aggregations?: Aggregation[];
-  group_by?: string[];
-  order_by?: string[];
-  limit?: number;
-  scenario?: string;
-  include_schema?: boolean;
-  include_sql?: boolean;
-};
-
-export type RapRow = {
-  topaz_id: number;
-  year: number;
-  band: number;
-  value: number;
-};
-
-export type AggregatedRapRow = {
-  year: number;
-  shrub: number;
-  tree: number;
-  coverage: number;
-};
-
-export type FetchRapOptions = {
-  mode: "hillslope" | "watershed";
-  runId: string;
-  topazId?: number;
-  year?: number;
-  include_schema?: boolean;
-  include_sql?: boolean;
-};
-
-export type FetchRapChoroplethOptions = {
-  runId: string;
-  band: number | number[];
-  year?: number | null;
-  include_schema?: boolean;
-  include_sql?: boolean;
-};
-
-export type RapChoroplethRow = {
-  wepp_id: number;
-  value: number;
-  shrub?: number;
-  tree?: number;
-};
-
-export type LanduseEntry = {
-  desc: string;
-  color: string;
-};
-
-export type LanduseMap = Record<number, LanduseEntry>;
-
-/**
- * Color modes for SBS (Soil Burn Severity) raster tiles and legend.
- * The backend is the source of truth; the frontend passes this value as
- * a query param to both the tile endpoint and GET /api/watershed/sbs/colormap.
- */
-export type SbsColorMode = "legacy" | "shift";
-
-export type SbsColormapEntry = {
-  /** SBS canonical class value (130 = Unburned … 133 = High) */
-  class_value: number;
-  /** Human-readable severity label */
-  label: string;
-  rgba: [number, number, number, number];
-  hex: string;
-};
-
-export type SbsColormapResponse = {
-  mode: SbsColorMode;
-  entries: SbsColormapEntry[];
-};
-
-export type RhessysSpatialLegendStop = {
-  value: number;
-  hex: string;
-};
-
-export type RhessysSpatialFile = {
-  filename: string;
-  name: string;
-  type: "continuous" | "categorical" | "stream";
-  min: number | null;
-  max: number | null;
-  unique_values: number[] | null;
-  group: string | null;
-  reversed: boolean;
-  legend: RhessysSpatialLegendStop[] | null;
-};
-
-export type RhessysSpatialListResponse = {
-  files: RhessysSpatialFile[];
-};
-
-export type RhessysOutputVariable = {
-  id: string;
-  label: string;
-  units: string;
-  filename: string;
-};
-
-export type RhessysOutputScenario = {
-  id: string;
-  label: string;
-  is_change: boolean;
-  variables: string[];
-};
-
-export type RhessysOutputValueRange = {
-  min: number;
-  max: number;
-};
-
-export type RhessysOutputListResponse = {
-  scenarios: RhessysOutputScenario[];
-  variables: RhessysOutputVariable[];
-  value_ranges?: Record<string, Record<string, RhessysOutputValueRange>>;
-};
-
-export type RhessysChoroplethRow = {
-  spatialId: number;
-  value: number;
-};
-
-export type SpatialScale = "hillslope" | "patch";
-
-export type FetchLanduseOptions = {
-  runId: string;
-  include_schema?: boolean;
-  include_sql?: boolean;
-  limit?: number;
-  scenario?: string;
-};
