+++ client/src/api/types/scenario.ts
